pyproject.toml
[tool.poetry]
name = "fgpyo"
version = "0.0.1"
description = "Python bioinformatics and genomics library"
authors = ["Nils Homer", "Tim Fennell", "Nathan Roach"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/fulcrumgenomics/fgpyo"
repository = "https://github.com/fulcrumgenomics/fgpyo"
keywords = ["bioinformatics"]
classifiers = [
	"Development Status :: 3 - Alpha",
    "Environment :: Console",
	"Intended Audience :: Developers",
	"Intended Audience :: Science/Research",
	"License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
	"Programming Language :: Python :: 3",
	"Topic :: Scientific/Engineering :: Bio-Informatics",
    "Topic :: Software Development :: Documentation",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
include = [
    "LICENSE",
]

[tool.poetry.dependencies]
python = ">=3.6.10"
attrs = ">=19.3.0"
samwell = ">=0.0.2"
typing_extensions = { version = ">=3.7.4", python = "<3.8" }  # Literal support
typing_inspect = { version = ">=0.3.1", python = "<3.8" }  # inspecting types

[tool.poetry.dev-dependencies]
pytest = ">=5.4.2"
mypy = ">=0.770"
flake8 = ">=3.8.1"
black = ">=19.10b0"
pytest-cov = ">=2.8.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
